Add solution for day 15
We try to fully explore the map while running the IntCode program. Then
we use BFS to find the shortest distance between the start and the goal.
Finally, to fill the maze with oxygen, we remove empty points from the
maze again in a breadth-first order, but we count each wave as a single
step.